Django-React-App/requirements.txt
asgiref==3.2.3
autopep8==1.5.7
beautifulsoup4==4.6.0
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
cryptography==3.4.7
defusedxml==0.6.0
Django==2.2.9
django-allauth==0.41.0
django-cogwheels==0.3
django-colorfield==0.1.15
django-cors-headers==3.7.0
django-filter==2.4.0
django-modelcluster==5.0.1
django-taggit==1.2.0
django-templated-mail==1.1.1
django-treebeard==4.3.1
django-wordpress-api==0.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.0.3
draftjs-exporter==2.1.7
html5lib==1.0.1
idna==2.10
iso8601==0.1.12
oauthlib==3.1.0
pbr==5.6.0
Pillow==6.2.2
pipenv==2020.11.15
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
selenium==3.141.0
simplejwt==2.0.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
stevedore==3.3.0
toml==0.10.2
typing==3.7.4.3
Unidecode==1.1.1
urllib3==1.26.2
virtualenv==16.7.9
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wagtail==2.7.1
wagtailmenus==3.0
wagtailstreamforms==3.8
webencodings==0.5.1
Willow==1.3
